[18] fixed connection error in case of exception and refactoring
