+++ src/main/kotlin/com/github/idell/gitlabconnect/ui/settings/CheckConnectionActionListener.kt
@@ -1,5 +1,6 @@
 package com.github.idell.gitlabconnect.ui.settings
 
+import com.github.idell.gitlabconnect.exception.GitlabConnectException
 import com.github.idell.gitlabconnect.services.gitlab.GitlabTestService
 import com.github.idell.gitlabconnect.storage.TokenConfiguration
 import java.awt.event.ActionEvent
@@ -10,5 +11,11 @@ class CheckConnectionActionListener(
     private val callback: (switch: Boolean) -> Unit
 ) : ActionListener {
 
-    override fun actionPerformed(e: ActionEvent?) = callback.invoke(GitlabTestService().test(getToken.invoke()))
+    override fun actionPerformed(e: ActionEvent?) {
+        try {
+            callback.invoke(GitlabTestService().test(getToken.invoke()))
+        } catch (e : GitlabConnectException){
+            callback.invoke(false)
+        }
+    }
 }
